+++ R/rcmd.R
@@ -21,6 +21,10 @@
 rcmd_build_tools <- function(..., env = character(), required = TRUE, quiet = FALSE) {
   env <- c(callr::rcmd_safe_env(), env)
 
+  if (!quiet) {
+    cli::cat_rule(paste0("R CMD ", ..1), col = "cyan")
+  }
+
   warn_for_potential_errors()
 
   callback <- if (cli::is_dynamic_tty()) {
@@ -29,16 +33,21 @@ rcmd_build_tools <- function(..., env = character(), required = TRUE, quiet = FA
     simple_callback(quiet)
   }
 
-  res <- with_build_tools({
-    withCallingHandlers(
-      callr::rcmd_safe(..., env = env, spinner = FALSE, show = FALSE,
-        echo = FALSE, block_callback = callback, stderr = "2>&1"),
-      error = function(e) {
-        if (!quiet) e$echo <- TRUE
-        asNamespace("callr")$err$throw(e)
-      }
-    )
-  }, required = required)
+  res <- with_build_tools(
+    {
+      withCallingHandlers(
+        callr::rcmd_safe(...,
+          env = env, spinner = FALSE, show = FALSE,
+          echo = FALSE, block_callback = callback, stderr = "2>&1"
+        ),
+        error = function(e) {
+          if (!quiet) e$echo <- TRUE
+          asNamespace("callr")$err$throw(e)
+        }
+      )
+    },
+    required = required
+  )
 
   msg_for_long_paths(res)
 
@@ -47,19 +56,21 @@ rcmd_build_tools <- function(..., env = character(), required = TRUE, quiet = FA
 
 msg_for_long_paths <- function(output) {
   if (is_windows() &&
-      any(grepl("over-long path length", output$stdout))) {
+    any(grepl("over-long path length", output$stdout))) {
     message(
       "\nIt seems that this package contains files with very long paths.\n",
       "This is not supported on most Windows versions. Please contact the\n",
       "package authors and tell them about this. See this GitHub issue\n",
-      "for more details: https://github.com/r-lib/remotes/issues/84\n")
+      "for more details: https://github.com/r-lib/remotes/issues/84\n"
+    )
   }
 }
 
 warn_for_potential_errors <- function() {
   if (is_windows() && grepl(" ", R.home()) &&
-      getRversion() <= "3.4.2") {
-    warning(immediate. = TRUE,
+    getRversion() <= "3.4.2") {
+    warning(
+      immediate. = TRUE,
       "\n!!! Building will probably fail!\n",
       "This version of R has trouble with building packages if\n",
       "the R HOME directory (currently '", R.home(), "')\n",
@@ -68,6 +79,7 @@ warn_for_potential_errors <- function() {
       "- installing it into a path without a space, or\n",
       "- creating a drive letter for R HOME via the `subst` windows command, and\n",
       "  starting R from the new drive.\n",
-      "See also https://github.com/r-lib/remotes/issues/98\n")
+      "See also https://github.com/r-lib/remotes/issues/98\n"
+    )
   }
 }